added cl warning lvl to build.bat

+++ include/common.h
@@ -17,7 +17,7 @@ typedef signed long long int s64;
 typedef float f32;
 typedef double f64;
 
-__declspec(dllexport) int StringLen(char *s);
+__declspec(dllexport) long long StringLen(char *s);
 __declspec(dllexport) bool StringCmp(char *s1, char *s2);
 __declspec(dllexport) char *StringConnect(char *s1, char *s2);
